==> src/age/00_QC_Luise.Rmd <==
---
title: "00_QC_Luise"
author: "Sera Choi"
date: "14/05/2021"
output: html_document
---

## Quality Control of Luise's dataset

* Assesses the quality of Luise's dataset of endothelial cells and pericytes 
in BA4 region of the human brain. 
* Subset of ECs and pericytes in BA4 contains 1,069 cells 
* The data has previously undergone rigorous quality control, but this script aims to 
double check and visualise some QC plots.

```{r, message=FALSE, warning=FALSE}
# Load library
library(Seurat)
library(ggplot2)
library(dplyr)
library(here)
library("scales")

mypal <- pal_npg("nrc", alpha = 0.7)(10)
mypal2 <-pal_tron("legacy", alpha = 0.7)(7)
mypal3 <- pal_lancet("lanonc", alpha = 0.7)(9)
mypal4 <- pal_simpsons(palette = c("springfield"), alpha = 0.7)(16)
mypal5 <- pal_rickandmorty(palette = c("schwifty"), alpha = 0.7)(6)
mypal6 <- pal_futurama(palette = c("planetexpress"), alpha = 0.7)(5)
mypal7 <- pal_startrek(palette = c("uniform"), alpha = 0.7)(5)
mycoloursP<- c(mypal, mypal2, mypal3, mypal4, mypal5, mypal6, mypal7)
show_col(mycoloursP, labels =F)

# Load data
ecpc <- readRDS(here("data/processed/ecpc.RDS")) 
```

### Proportion of mito transcript 
When Luise quality controlled the dataset, spliced and unspliced transcripts were QCed using different threshold values. Since no mito genes are present in unspliced, upper threshold for mito genes was set for only the spliced genes. Then counts for unspliced and spliced were combined. Therefore, the current values in 'percent.mt' reflect mito genes for only spliced genes. Here, I re-calculated the mito percentage and compared with the old. The mito percentage decreases because now, there are more non-mito genes (unspliced), reducing the proportion of mito genes present. ecpc00.RDS contains revised mito percentage.    


```{r}
ecpc[['percent.mt_new']]<- PercentageFeatureSet(ecpc, pattern = "^MT-")
VlnPlot(ecpc, features = c("percent.mt", "percent.mt_new"), ncol = 2)

# Update seurat object with re-calculated mito values 
ecpc <- readRDS(here("data/processed/ecpc.RDS")) 
ecpc[['percent.mt']]<- PercentageFeatureSet(ecpc, pattern = "^MT-")
saveRDS(ecpc, "/Users/serachoi/Documents/Edinburgh/MScProject/Project/data/processed/ecpc00.RDS")
ecpc <- readRDS(here("data/processed/ecpc00.RDS")) 
```


### QC plots 
QC aims to remove any low quality libraries. QC assesses:

* number of expressed genes per cell (nFeature_RNA)
* number of molecules per cell (nCount_RNA) - count depth 
* proportion of mito genes per cell (percent.mt)

Often cells with low count depth, few detected genes and high mito transcript represent dead or damaged cells. In contrast, high count depth and large number of detected genes represent doublets. Therefore, thresholds are set for:

* lower and upper thresholds for number of expressed genes per cell (nFeature_RNA)
* lower and upper thresholds for count depth (nCount_RNA)
* upper thresholds for mito genes 

Luise's thresholds are:

* number of expressed genes: upper=8000, lower=10
* number of molecules: upper=30000, lower= 100
* mito genes upper: 23%

Here, I plotted violin plot for three covariates and created a table showing min and max values for each covariate in the dataset. The minimum and maximum values in the dataset agrees with the thresholds. I created plots at two endothelial population level and at sample level. 

```{r, fig.width=15, fig.height=5}
# At cell annotation level
VlnPlot(ecpc, 
        features = c("nFeature_RNA","nCount_RNA", "percent.mt"), 
        ncol = 3)

# At sample level 
pdf("/Users/serachoi/Documents/Edinburgh/MScProject/Project/outs/method/SeekerQC.pdf", width = 15, height = 5 )
VlnPlot(ecpc, 
        features = c("nFeature_RNA","nCount_RNA", "percent.mt"), 
        ncol = 3, 
        group.by = "BBN",
        pt.size = FALSE,
        cols = mycoloursP[1:20])
dev.off()

# Create table containing min and max values of count depth, expressed genes and mito %
metadata <- ecpc@meta.data #metadata
minimum <- c(min(metadata$nFeature_RNA), min(metadata$nCount_RNA), min(metadata$percent.mt))
maximum <- c(max(metadata$nFeature_RNA), max(metadata$nCount_RNA), max(metadata$percent.mt))
min_max_df <- data.frame(minimum, maximum)
rownames(min_max_df) <- c("nFeature_RNA", "nCount_RNA", "percent.mito")
min_max_df

```

FeatureScater is used to visualise feature-feature relationships. Pearson correlation between two features is displayed at the top.

**Plot 1: number of molecules vs. mito percent** 

* Cells with low number of molecules have high mito % because with smaller number of non-mito transcripts, presence of mito transcript affects the proportion 
* Cells with high number of molecules generally have low mito % because high non-mito trasncript masks away mito transcripts 
* Almost no correlation 

**Plot 2: number of molecules vs. number of unique genes**

* Positive correlation between total number of molecules and genes per cell because if a cell has many molecules, more different genes are expressed 
* As total number of molcules in the cell increases, the number of unique gene increases at decreasing rate because 3 molecules of the same gene is just one gene. 

```{r, fig.width=10, fig.height=4, fig.fullwidth=TRUE}
plot1 <- FeatureScatter(ecpc, feature1 = "nCount_RNA", feature2 = "percent.mt", group.by = "BBN")
plot2 <- FeatureScatter(ecpc, feature1 = "nCount_RNA", feature2 = "nFeature_RNA", group.by = "BBN")
plot1 + plot2
```

# Conlcusion 

* Quality of cells in Luise's dataset appear good 
* No major cause of concern.
