==> src/age/02_DimensionReduction_Luise.Rmd <==
---
title: "02_DimensionReduction_Luise"
author: "Sera Choi"
date: "19/05/2021"
output: html_document
---

```{r, message=FALSE, warning=FALSE}
# Load library
library(Seurat)
library(ggplot2)
library(dplyr)
library(here)

# Load data
ecpc <- readRDS(here("data/processed/ecpc01.RDS")) 
```

# Scaling data 
After selecting top 2000 HVGs, linear transformation ('scaling') is applied. In Seurat, the ScaleData() function:

* shifts expression of each gene so that mean expression across cells is 0
* shifts expression of each gene so that the variance across cells is 1
* results of this is stored in ecpc[["RNA"]]@scale.data 

```{r, message=FALSE}
all.genes <- rownames(ecpc)
ecpc <- ScaleData(ecpc, features = all.genes)
```

# Linear Dimensional reduction (PCA)
PCA was performed on the scaled data. In Seurat, PCA is run on top 2000 HVGs previously identified during feature selection. One PC captures variance across all cells, with top PCs capturing dominant factors that give rise to heterogeneity in the data. 

When performing PCA, we assume that "biological processes affect multiple genes in a coordinated manner" (Bioconductor). Therefore, top PCs represent biological variation that cause correlated behaviour of many genes. In comparison, random technical noise tend to affect each gene independently and therefore is represented by later PCs. This is why we concentrate on top few PCs to retain in downstream analysis.   

Below shows top 5 PCs with positive and negative scores reflecting the positive and negative correlation of genes with the PCs. If a gene has PC score = 0, the gene doesn't play any role in explaining variation on that particular PC. The output below only displays top 5 positive and negative genes per PC, but that list continues to include all 2000 genes (unless some genes have PC score = 0). 

```{r, message=FALSE}
ecpc <- RunPCA(ecpc, features= VariableFeatures(ecpc))
print(ecpc[["pca"]], dims = 1:5, nfeatures = 5)
```

 The magnitude of scores reflects how much it correlates. 

```{r}
head(Embeddings(ecpc, reduction = "pca")[, 1:5])
```

## Visualisation

Clearly, the two cell populations are separating by principal component 1. 

```{r}
DimPlot(ecpc, reduction='pca')
```

# PCA Heatmap

Here, I plotted heatmap of dimensional reduction, with cells on column and genes on row. Both cells and genes are sorted by their principal component scores. Principal component score increases from negative (ELOVL7) to 0 to positive (PRKG1). Extreme cells are placed on the two ends on the plot. For each PC, genes with top positive and negative scores are plotted on row while, extreme cells are plotted on column.

```{r}
DimHeatmap(ecpc, dims = 1, cells = 10, balanced = TRUE)
```

Top PCs represent dominant factors that give rise to heterogeneity in the dataset. Therefore, at PC1, a distinct split can be seen of cells with up- and down-regulated expression of certain genes. 

Elbow plot is used to determine the number of PCs to retain for downstream analysis. The plot shows the percentage of variance explained as a function of increasing PC. The threshold will be chosen at a point in which adding another PC does not significantly increase the variance. To begin with I chose 25. 

```{r}
ElbowPlot(ecpc, ndims = 50)
```

```{r}
# Update seurat object 
saveRDS(ecpc, "/Users/serachoi/Documents/Edinburgh/MScProject/Project/data/processed/ecpc02.RDS")
```
