==> src/age/04_annotation_Luise.Rmd <==
---
title: "04_annotation_Luise"
author: "Sera Choi"
date: "24/05/2021"
output: html_document
---

```{r, message=FALSE, warning=FALSE}
# Load library
library(Seurat)
library(ggplot2)
library(dplyr)
library(here)
library("scales")
library(ggsci)
library(gtools)

# Load data
ecpc <- readRDS(here("/Users/serachoi/Documents/Edinburgh/MScProject/Project/data/processed/ecpc03.RDS")) 
```

# Annotation of clusters 

This section tries to identify clusters at resolution ranging from 0.2-0.5. We hypothesise that some of those cells are ECs, pericytes and smooth muscle cells. Gene markers for ECs and pericytes were found in literatures (Luise and 'Reconstruction of the human blood–brain barrier in vitro reveals a pathogenic mechanism of APOE4 in pericytes', 2020) : 

* Endothelial:CLDN5, VWF, FLT1, ANO2, ELOVL7 and ABCB1
* Pericytes: PDGFRB, NOTCH3, PRKG1, LAMA2, GPC5, PDE7B, and DLC1
* SMCs: MYH11, TAGLN, ACTA2, LPP and MYL9

## Endothelial 

* Concentrates around upper population of cells. 

```{r, fig.width= 15, fig.height= 13}
FeaturePlot(ecpc, features= c("CLDN5", 
                               "VWF", 
                               "FLT1", 
                               "ANO2", 
                               "ELOVL7", 
                               "ABCB1"))
```

## Pericyte

* Concnetrated in lower population of cells. 

```{r, fig.width= 15, fig.height= 13}
FeaturePlot(ecpc, features= c("PDGFRB", 
                              "NOTCH3", 
                              "PRKG1", 
                              "LAMA2", 
                              "GPC5", 
                              "PDE7B"))
```

## Smooth Muscle cell (vascular cells)

* Highlights a small population off pericyte-like population

```{r, fig.width= 15, fig.height= 13}
FeaturePlot(ecpc, features= c("MYH11", 
                              "TAGLN", 
                              "ACTA2", 
                              "MYL9"))
```

## Annotation at resolution 0.25

Using canonical gene markers for ECs, pericytes and smooth muscle cells, we could annotate cluster 0, 1 and 3. Although cells in cluster 2 appear to express EC gene markers, the cluster seems interesting as it clusters away from the main EC population. 

```{r}
ecpc <- FindClusters(ecpc, resolution = 0.25)
ecpc <- RunUMAP(ecpc, dims = 1:25)
DimPlot(ecpc, reduction="umap", label=TRUE, label.size = 8) + NoLegend()
```

# Feature plots with identified markers 

Here, the main aim is to identify characteristics of cells in cluster 2. However, I also performed DE analysis on clusters 0, 1, and 3 just to double check the cell identity.  

## Gene Marker Identification

Seurat uses function FindAllMarkers() to identify positive and negative gene markers of one cluster, compared to all other cells. So, if a gene in cluster 0 has avg.log2FC of 2.0 with highly significant adjusted p-value, then, cells in cluster 0 express the gene on average 2-fold more than cells in cluster 1 and 2.    

* min.pct : if min.pct = 0.25, genes that are expressed in 25% of cells in either of the two populations are only tested. This reduces computational cost as irrelavant genes are not tested.
* logfc.threshold : only those genes with average log X-fold difference between two groups of cells are tested.  
* only positive gene markers are returned. 
* By default, Wilcoxon Rank Sum test is used but here, I chose to use MAST because literatures have identified MAST to be best performing single-cell DE testing method. 

```{r, message=FALSE}
# Set cell identity class to clustering resolution 0.2
Idents(ecpc) <- "RNA_snn_res.0.25"

# Finds gene markers in each cluster compared to all other cells 
ecpc.markers <- FindAllMarkers(ecpc, 
                               only.pos = TRUE, 
                               min.pct = 0.25, 
                               logfc.threshold = 0.25,
                               test.use = "MAST")
```

The results below show 10 gene markers for each cluster ordered with highest log2FC: 

* unadjusted p-value 
* log2 fold-change of average expression between two groups. Positive value indicates that cells in first group express the feature more 
* pct1: percentage of cells that express the feature in group 1
* pct2: percentage of cells that express the feature in group 2
* adjusted p-value after Bonferroni correction 

Since only positive gene markers are reported here, pct.1 > pct.2. At this point, you can further filter gene markers based on 4 parameters.

* pct1 > 0.25% is quite a normal threshold. 
* pct2 depends on your dataset. If your dataset contains similar cells, pct2 < 0.25 is suitable whereas, if dataset contains heterogenous population of cells, pct2 < 0.6 is suitable. 
* adjusted p-value < 0.05 for significant results 
* log2fc should be considered because if all cells are expressing very poorly but only slightly different, those genes are not significant. 

```{r, message=FALSE}
top10_markers <- ecpc.markers %>% group_by(cluster) %>% top_n(n = 10, wt = avg_log2FC)
top10_markers
```

## Cluster 0 

* Known gene markers for ECs seen (ABCB1 and THSD4): further confirms that cluster 0 is ECs

```{r}
# This function takes result from DE analysis and creates feature plots 
DEG_feature_plot_func <- function(seur_object,
                                  markers){
  
  # Converts data.frame into string of DEGs
  gene.markers <- as.character(as.matrix(markers))
  
  # Create feature plot using makers 
  feat_plot <- FeaturePlot(seur_object, 
                           features = c(gene.markers))
  print(feat_plot)
}
```


```{r, fig.width= 15, fig.height= 13}
# Cluster 0 annotation with 10 identified markers 
clust0_markers <- top10_markers[1:10, 'gene']
DEG_feature_plot_func(ecpc, clust0_markers)
```

## Cluster 1 gene marker 

* Known pericyte markers seen (PRKG1, GPC5, DLC1)

```{r, fig.width= 15, fig.height= 13}
# Cluster 1 top 10 gene markers 
clust1_markers <- top10_markers[11:20, 'gene']
DEG_feature_plot_func(ecpc, clust1_markers)
```

## Cluster 2 gene marker

* See later thorough analysis on this cluter. 

```{r, fig.width= 15, fig.height= 13}
# Cluster 2 top 10 gene markers 
clust2_markers <- top10_markers[21:30, 'gene']
DEG_feature_plot_func(ecpc, clust2_markers)
```

## Cluster 3 gene marker 

* SMC gene markers identified. 

```{r, fig.width= 15, fig.height= 13}
# Cluster 3 top 10 gene markers 
clust3_markers <- top10_markers[31:40, 'gene']
DEG_feature_plot_func(ecpc, clust3_markers)
```

So far, cluster 0, 1, and 3 have been identified to be ECs, pericytes and SMCs. Cluster 2 appears to be ECs as well but still interesting. 

# Annotation at resolution = 0.5

Here, I attempt to identify the two sub-clusters in endothelial population. Initially I thought it could be different vessel-types (arterioles, venuoles and capillary), but it turns out that it was quality. This is also true for pericyte sub-population 

```{r}
ecpc <- FindClusters(ecpc, resolution = 0.5)
ecpc <- RunUMAP(ecpc, dims = 1:25)
DimPlot(ecpc, reduction="umap", label=TRUE, label.size = 8) + NoLegend()

FeaturePlot(ecpc, features = c("nFeature_RNA", "nCount_RNA"))
```

# Conclusion

Since sub-clusters in EC population is due to quality, clusters will be annotated at resolution 0.25. The Endothelial_2 population is still unknown but next section will try address this.  

```{r}
Idents(ecpc) <- "RNA_snn_res.0.25"
new.cluster.ids <- c("Endothelial_1", "Pericyte", "Endothelial_2", "SMC")
names(new.cluster.ids) <- levels(ecpc)
ecpc <- RenameIdents(ecpc, new.cluster.ids)
DimPlot(ecpc, reduction = "umap", label = TRUE, label.size = 8) + NoLegend()
```

```{r}
# Save annotated RDS 
saveRDS(ecpc, "/Users/serachoi/Documents/Edinburgh/MScProject/Project/data/processed/ecpc04.RDS")
```
