==> src/age/08_Yang.Rmd <==
---
title: "08_Yang"
author: "Sera Choi"
date: "02/07/2021"
output: html_document
---

```{r, message=FALSE, warning=FALSE}
# Load library
library(Seurat)
library(ggplot2)
library(dplyr)
library(here)
library("scales")
library(ggsci)
library(gtools)
library(future)

mypal <- pal_npg("nrc", alpha = 0.7)(10)
mypal2 <-pal_tron("legacy", alpha = 0.7)(7)
mypal3 <- pal_lancet("lanonc", alpha = 0.7)(9)
mypal4 <- pal_simpsons(palette = c("springfield"), alpha = 0.7)(16)
mypal5 <- pal_rickandmorty(palette = c("schwifty"), alpha = 0.7)(6)
mypal6 <- pal_futurama(palette = c("planetexpress"), alpha = 0.7)(5)
mypal7 <- pal_startrek(palette = c("uniform"), alpha = 0.7)(5)
mycoloursP<- c(mypal, mypal2, mypal3, mypal4, mypal5, mypal6, mypal7)
show_col(mycoloursP, labels =F)

options(future.globals.maxSize = 50000*1024^2)
```

# Yang et al (GSE163577)

The purpose of this section is to extract ECs from the snRNA-seq data generated by Yang et al. 

* snRNA-seq: 10x version 3
* Superior frontal cortex tissue
* 4 individuals aged 58(M), 77(M), 79(F), 82(M)

```{r}
yang <- readRDS("/Users/serachoi/Documents/Edinburgh/MScProject/Project/data/Yang_et_al.RDS")
```

## Quality Control

Following Yang et al, nuclei with percent.mito > 5%, features < 200 and features > 5,000 were removed. 

```{r}
yang[["percent.mt"]] <- PercentageFeatureSet(yang, pattern = "^MT-")
VlnPlot(yang, features = c("nFeature_RNA", "nCount_RNA", "percent.mt"), ncol = 3, pt.size = 0, cols = mycoloursP[20:30])

yang <- subset(yang, subset = nFeature_RNA > 200 & nFeature_RNA < 5000 & percent.mt < 5)
pdf("/Users/serachoi/Documents/Edinburgh/MScProject/Project/outs/method/YangQC.pdf", width = 15, height = 5 )
VlnPlot(yang, features = c("nFeature_RNA", "nCount_RNA", "percent.mt"), ncol = 3, pt.size = 0, cols = mycoloursP[20:30])
dev.off()
plot1 <- FeatureScatter(yang, feature1 = "nCount_RNA", feature2 = "percent.mt")
plot2 <- FeatureScatter(yang, feature1 = "nCount_RNA", feature2 = "nFeature_RNA")
plot1 + plot2
```

## SCTransform

The study used SCTtransform normalisation and integration. Requires further integration. 

```{r}
yang <- SCTransform(yang, vars.to.regress = "percent.mt", verbose = TRUE)
yang <- RunPCA(yang, verbose = FALSE)

yang <- RunUMAP(yang, dims = 1:30, verbose = FALSE)
pdf("/Users/serachoi/Documents/Edinburgh/MScProject/Project/outs/age/yang_inte1", width = 8, height = 5)
DimPlot(yang, reduction="umap", cols = mycoloursP[1:20])
dev.off()
```


## Integration 

```{r, message=FALSE, warning=FALSE}
# Split Seurat object into 4 objects corresponding to each sample 
yang.list <- SplitObject(yang, split.by = "orig.ident")

# Normalise and identify HVGs for each dataset independently
yang.list <- lapply(X = yang.list, FUN = function(x) {
    x <- NormalizeData(x)
    x <- FindVariableFeatures(x, selection.method = "vst", nfeatures = 2000)
})

# select features that are repeatedly variable across datasets for integration
features <- SelectIntegrationFeatures(object.list = yang.list)
yang.anchors <- FindIntegrationAnchors(object.list = yang.list, anchor.features = features, dims = 1:20)
yang.combined <- IntegrateData(anchorset = yang.anchors, dims=1:20)

DefaultAssay(yang.combined) <- "integrated" 
```

## Feature selection, dimensional reduction and visualisation

```{r, fig.width=15, fig.height=10}
# Run the standard workflow for visualization and clustering
# Feature Selection and Scaling 
yang.combined <- FindVariableFeatures(yang.combined, selection.method = "vst", nfeatures = 2000)
all.genes <- rownames(yang.combined)
yang.combined <- ScaleData(yang.combined, features = all.genes)

# Linear dimensional reduction 
yang.combined <- RunPCA(yang.combined, features = VariableFeatures(object = yang.combined))
ElbowPlot(yang.combined, ndims = 50)

# Visualisation
yang.combined <- RunUMAP(yang.combined, dims = 1:20)
```

```{r}
pdf("/Users/serachoi/Documents/Edinburgh/MScProject/Project/outs/age/yang_inte2", width = 8, height = 5)
DimPlot(yang.combined, reduction = "umap", group.by = "orig.ident", cols = mycoloursP[1:20])
dev.off()
```

# Cluster

```{r}
#Generate DimPlot of all tested clustering resolutions in metadata
# requires gtools library and Seurat
plot_list_func <- function(seur_obj,
                           col_pattern,
                           clust_lab = TRUE,
                           label_size = 8,
                           num_col = 4){
  # finds column with additional resolution clusterings
  extr_res_col <- grep(pattern = col_pattern, names(seur_obj@meta.data))
  # obtain column names
  res_names <- names(seur_obj@meta.data[extr_res_col])
  
  # gtools function, sorts gene_names alphanumeric:
  res_names <- mixedsort(res_names) 
  plot_l <-list()
  
  for(i in 1: length(res_names)){
  dim_plot <- DimPlot(seur_obj, 
                      reduction = "umap", 
                      group.by = res_names[i],
                      label = clust_lab,
                      label.size = label_size) + NoLegend()
  print(dim_plot)
 }
}

yang.combined <- FindNeighbors(yang.combined, dims = 1:20)
# test different resolutions for clustering
yang.combined <- FindClusters(yang.combined, resolution = c(0.01, 0.05, 
                                        seq(from = 0.1, to = 1, by = 0.1)))
# non-linear reduction
yang.combined <- RunUMAP(yang.combined, dims = 1:20)

plot_list_func(seur_obj = yang.combined,
               col_pattern = "integrated_snn_res.")
```

# Feature Plots 

Yang et al were able to identify arterial, capillary and venous ECs within the brain. It appears that cluster 1 at resolution 0.01 is the endothelial cells. 

```{r}
# Arterial
FeaturePlot(yang.combined, features = c('ARL15', 'VEGFC', 'MECOM', 'THSD7A'))

# Capillary
FeaturePlot(yang.combined, features= c('EMP1','FLT1','ATP10A','SLC7A5'))

# Vein
FeaturePlot(yang.combined, features= c('TSHZ2','AFF3','LINC02147','SNTG2'))
```

## Subset for ECs 

There are 19,655 BECs. 

```{r}
Idents(yang.combined) <- "integrated_snn_res.0.01"
pdf("/Users/serachoi/Documents/Edinburgh/MScProject/Project/outs/age/yang_EC", width = 10, height = 5)
FeaturePlot(yang.combined, features= c("CLDN5","ABCB1"), label = T, label.size = 8)
dev.off()
table(Idents(yang.combined))

# Subset 
yang_ec <- subset(yang.combined, idents = 1)
saveRDS(yang_ec, "/Users/serachoi/Documents/Edinburgh/MScProject/Project/data/processed/yang_ec.RDS")
```

# Analysing ECs from Yang et al

```{r}
ec <- readRDS("/Users/serachoi/Documents/Edinburgh/MScProject/Project/data/processed/yang_ec.RDS")

# Feature selection 
ec <- FindVariableFeatures(ec, selection.method = "vst", nfeatures = 2000)

# Scale data
all.genes <- rownames(ec)
ec <- ScaleData(ec, features = all.genes)

# Dimensionality reduction
ec <- RunPCA(ec, features = VariableFeatures(object = ec))
ElbowPlot(ec, ndims=50)

# Visualisation
ec <- RunUMAP(ec, dims = 1:20)
DimPlot(ec, reduction="umap")

# Clustering at different resolutions
ec <- FindNeighbors(ec, dims = 1:20)
ec <- FindClusters(ec, resolution = c(0.01, 0.05, seq(from = 0.1, to = 1, by = 0.1)))
ec <- RunUMAP(ec, dims = 1:20)
plot_list_func(seur_obj = ec, col_pattern = "integrated_snn_res.")
```

### BEC arterial, capillary and vein markers 

Using gene markers Yang et al used, identified capillary BECs. 

```{r, fig.width=10, fig.height=5}
Idents(ec) <- "integrated_snn_res.0.1"
DefaultAssay(ec) <- "RNA"
# Arterial
pdf("/Users/serachoi/Documents/Edinburgh/MScProject/Project/outs/age/yang_artery", width = 10, height = 5)
FeaturePlot(ec, features = c('ARL15', 'VEGFC'), label = T, label.size = 8, max.cutoff = 10)
dev.off()
# Capillary
pdf("/Users/serachoi/Documents/Edinburgh/MScProject/Project/outs/age/yang_cap", width = 10, height = 5)
FeaturePlot(ec, features= c('SLC7A1','ABCB1'), label = T, label.size = 8, max.cutoff = 10)
dev.off()
# Vein
pdf("/Users/serachoi/Documents/Edinburgh/MScProject/Project/outs/age/yang_vein", width = 10, height = 5)
FeaturePlot(ec, features= c('TSHZ2','AFF3'), label = T, label.size = 8, max.cutoff = 10)
dev.off()
```

### Subset capillary ECs 

```{r}
Idents(ec) <- "integrated_snn_res.0.1"
id <- c("Capillary","Venous","Arterial")
names(id) <- levels(ec)
ec <- RenameIdents(ec, id)
table(Idents(ec))
DimPlot(ec, reduction = "umap", label = TRUE, label.size = 6) +NoLegend()

capillary_ec <- subset(ec, idents = "Capillary")
saveRDS(capillary_ec, "/Users/serachoi/Documents/Edinburgh/MScProject/Project/data/processed/yang_capillary_ec.RDS")
```
